Add search teams on org API
- Add the API to search for teams on a organization by the
`/orgs/{org}/teams/search` API of Gitea.
- The response body of the API is a bit weird because it the JSON can
return three fields "data", "error", "ok", first check if there's a
general HTTP error. Otherwise when ok is set to true,
return a error with the given error message of the error field. When
that ok is set to true, simply return the data.